Split types.hpp into fd.hpp, process.hpp cpu.hpp
