+++ include/lo2s/monitor/abstract_process_monitor.hpp
@@ -27,7 +27,7 @@ extern "C"
 #include <sys/types.h>
 }
 
-#include <lo2s/types.hpp>
+#include <lo2s/types/process.hpp>
 
 namespace lo2s
 {
