+++ src/types/process.cpp
@@ -20,7 +20,8 @@
  */
 
 #include <lo2s/execution_scope.hpp>
-#include <lo2s/types.hpp>
+#include <lo2s/types/process.hpp>
+
 namespace lo2s
 {
 
@@ -51,20 +52,4 @@ ExecutionScope Cpu::as_scope() const
 {
     return ExecutionScope(*this);
 }
-
-bool operator==(const WeakFd& lhs, const Fd& rhs)
-{
-    return lhs.fd_ == rhs.fd_;
-}
-
-bool operator==(const Fd& lhs, const WeakFd& rhs)
-{
-    return lhs.fd_ == rhs.fd_;
-}
-
-Fd::operator WeakFd() const
-{
-    return WeakFd(fd_);
-}
-
 } // namespace lo2s
